pyproject.toml
[project]
name = "linkedin-rpa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "ruff>=0.14.5",
    "selenium>=4.38.0",
    "claude-agent-sdk>=0.0.14",
    "pypdf>=5.0.0",
]
